Add --sign-commit option to build-chunked-oci
This adds a new --sign-commit CLI option to `rpm-ostree experimental
compose build-chunked-oci` which signs the embedded ostree commit.

With this, it is possible (with some care) to layer a signed bootc
image that uses old-school signed ostree/composefs support.

Note, this relies on the double free fix for ostree_sign_set_sk() in
https://github.com/ostreedev/ostree/pull/3527 to work.