build-chunked-oci: Correctly label files
Running repo.write_dfd_to_mtree() on the subdirs individually will
cause the modifier to label files incorrectly. For example when
calling it on the "etc" subdirectory, it will try to label /etc/foo as
/foo.

Instead we relabel the entire image, but use a filter to strip out
/sysroot content.

This is slightly complicated by the tests also wanting to use a
modifiers, so we add the no_attr argument to
generate_commit_from_rootfs() that does what the tests need.