treefile: Add `no-initramfs:` knob
In the work to rebase FCOS/RHCOS to fedora-bootc/rhel-bootc, we want to
run postprocess scripts for CoreOS specific changes to the rootfs, but
we want changes made in those postprocess scripts to affect the
initramfs generation too.

The new experimental `no-initramfs` knob will allow us to disable the
generation of the initramfs via dracut in the initial compose of the
rootfs. Then we can run our postprocess scripts on the rootfs and later
run dracut to generate the initramfs.

This came from discussion in https://gitlab.com/fedora/bootc/base-images/-/merge_requests/315
where it was decided to not add a `--postprocess` argument to bootc-base-imagectl.

Co-authored-by: Jonathan Lebon <jonathan@jlebon.com>